Move table preview app to its own package
