+++ packages/table-dev-app/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@blueprintjs/table-dev-app",
+  "version": "0.1.0",
+  "description": "Dev application for @blueprintjs/table",
+  "private": true,
+  "scripts": {
+    "bundle": "webpack --config ./webpack.config.js",
+    "clean": "rm -rf dist/*",
+    "dev": "webpack-dev-server --config ./webpack.config.js",
+    "dist": "NODE_ENV=production yarn bundle",
+    "lint": "npm-run-all -p lint:stylelint lint:tslint",
+    "lint:stylelint": "stylelint -s scss 'src/**/*.scss'",
+    "lint:tslint": "tslint '{src,test}/**/*.{ts,tsx}' -c ../tslint-config",
+    "lint-fix": "yarn lint:tslint --fix",
+    "test": "exit 0",
+    "verify": "npm-run-all -p dist lint"
+  },
+  "dependencies": {
+    "@blueprintjs/core": "*",
+    "@blueprintjs/table": "*",
+    "classnames": "^2.2.5",
+    "dom4": "^1.8.3",
+    "normalize.css": "~4.1.1",
+    "pure-render-decorator": "^1.2.1",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1"
+  },
+  "devDependencies": {
+    "@blueprintjs/webpack-build-scripts": "^0.1.0",
+    "npm-run-all": "^4.1.2",
+    "tslint": "^5.8.0",
+    "stylelint": "^8.2.0",
+    "webpack-dev-server": "^2.9.3",
+    "webpack": "^3.8.1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:palantir/blueprint.git"
+  },
+  "author": "Palantir Technologies",
+  "license": "Apache-2.0"
+}
