+++ packages/table-dev-app/src/index.scss
@@ -3,7 +3,7 @@
 
 @import "~normalize.css/normalize.css";
 @import "~@blueprintjs/core/dist/blueprint.css";
-@import "../src/table.scss";
+@import "~@blueprintjs/table/src/table.scss";
 
 // TODO: convert below CSS to SCSS
