+++ packages/table-dev-app/src/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../../config/tsconfig.base.json",
+    "compilerOptions": {
+        "declaration": false,
+        "lib": ["dom", "es5", "es6"]
+    }
+}
