+++ packages/table/.npmignore
@@ -1,5 +1,3 @@
 config/
 coverage/
-preview/
-dist/preview/
 test/
