--- packages/table/preview/tsconfig.json
@@ -1,6 +1,0 @@
-{
-    "extends": "../src/tsconfig",
-    "compilerOptions": {
-        "declaration": false
-    }
-}
